restructure the support model
* Splitted helpers up in separate files for improved clarity, and to
prepare for adding new helpers
* Removed in-library logging setup, as it could cause issues for
upstream users who wish to configure the library's logger
* Adapted tests

This is not thoroughly tested yet, even though the test suite passes.
There might still be a few missing imports or similar issues lurking
around that are not covered by unit tests.
