+++ src/pypdfium2/__init__.py
@@ -1,17 +1,21 @@
 # SPDX-FileCopyrightText: 2022 geisserml <geisserml@gmail.com>
 # SPDX-License-Identifier: Apache-2.0 OR BSD-3-Clause
 
-import logging
-from pypdfium2._logging import setup_logger
-logger = logging.getLogger(__name__)
-setup_logger(logger)
-
 import atexit
+import logging
 from pypdfium2 import _version
 from pypdfium2._pypdfium import *
-from pypdfium2._helpers import *
-from pypdfium2._constants import *
-from pypdfium2._exceptions import *
+from pypdfium2._helpers.constants import *
+from pypdfium2._helpers.error_handler import *
+from pypdfium2._helpers.exceptions import *
+from pypdfium2._helpers.opener import *
+from pypdfium2._helpers.page_renderer import *
+from pypdfium2._helpers.pdf_renderer import *
+from pypdfium2._helpers.toc import *
+from pypdfium2._helpers.utilities import *
+
+
+logger = logging.getLogger(__name__)
 
 __version__ = _version.V_PYPDFIUM2
 __pdfium_version__ = _version.V_LIBPDFIUM
