+++ src/pypdfium2/_helpers/__init__.py
@@ -1,9 +1,2 @@
 # SPDX-FileCopyrightText: 2022 geisserml <geisserml@gmail.com>
 # SPDX-License-Identifier: Apache-2.0 OR BSD-3-Clause
-
-import logging
-
-
-def setup_logger(logger):
-    handler = logging.StreamHandler()
-    logger.addHandler(handler)
