+++ src/pypdfium2/_helpers/exceptions.py
