+++ tests/conftest.py
@@ -1,7 +1,12 @@
 # SPDX-FileCopyrightText: 2022 geisserml <geisserml@gmail.com>
 # SPDX-License-Identifier: Apache-2.0 OR BSD-3-Clause
 
+import logging
 from pathlib import Path
+
+
+lib_logger = logging.getLogger('pypdfium2')
+lib_logger.addHandler(logging.StreamHandler())
 
 
 TestDir     = Path(__file__).parent.resolve()
